+++ src/MyAlgo2.java
@@ -43,7 +43,7 @@ public class MyAlgo2 {
             job.processingTime = Math.floor(job.processingTime/K);
             job.dueTime = job.dueTime/K;
         }
-        OurSchedule approximateScaledSchedule = this.getSchedule(scaledSchedule,0,0);
+        OurSchedule approximateScaledSchedule = this.getSchedule(scaledSchedule,0d,0);
         OurSchedule resultSchedule = new OurSchedule();
         for(Job scaledJob: approximateScaledSchedule){
             for(Job job: schedule){
@@ -59,7 +59,7 @@ public class MyAlgo2 {
     }
 
     // 3. The private function
-    private OurSchedule getSchedule(OurSchedule schedule, int timePassed, int level){
+    private OurSchedule getSchedule(OurSchedule schedule, Double timePassed, int level){
 
         // Step3.0 - return if same schedule exists
         OurSchedule memoizationCandidate = memoization.get(new Tuple(schedule, timePassed));
@@ -85,13 +85,13 @@ public class MyAlgo2 {
 //        }
 
         // Step3.3 - loop over all delta [0,N-kID]
-        int minimumTardiness = Integer.MAX_VALUE;
+        Double minimumTardiness = Double.MAX_VALUE;
         for (int delta=0; delta <= N-kId; delta++){
 
             // Step3.3.1 - Split into 3 branches
             OurSchedule jobsBranch1 = schedule.getSubset(0, kId-1).concatenate(schedule.getSubset(kId+1,kId+delta));
             OurSchedule jobsBranch2 = schedule.getSubset(0,kId+delta);
-            int ck = timePassed+jobsBranch2.getProcessingTime();
+            Double ck = timePassed+jobsBranch2.getProcessingTime();
             OurSchedule jobsBranch3 = schedule.getSubset(kId+delta+1, N);
 
             // Step 3.3.2 - Split into 3 branches
